+++ just_audio/android/src/main/java/com/ryanheise/just_audio/MainMethodCallHandler.java
@@ -34,7 +34,18 @@ public class MainMethodCallHandler implements MethodCallHandler {
                 break;
             }
             List<Object> rawAudioEffects = call.argument("androidAudioEffects");
-            players.put(id, new AudioPlayer(applicationContext, messenger, id, call.argument("audioLoadConfiguration"), rawAudioEffects, call.argument("androidOffloadSchedulingEnabled")));
+            players.put(
+                id,
+                new AudioPlayer(
+                    applicationContext,
+                    messenger,
+                    id,
+                    call.argument("audioLoadConfiguration"),
+                    rawAudioEffects,
+                    call.argument("androidOffloadSchedulingEnabled"),
+                    call.argument("userAgent")
+                )
+            );
             result.success(null);
             break;
         }