+++ just_audio_platform_interface/lib/just_audio_platform_interface.dart
@@ -397,6 +397,7 @@ class InitRequest {
   final List<AudioEffectMessage> androidAudioEffects;
   final List<AudioEffectMessage> darwinAudioEffects;
   final bool? androidOffloadSchedulingEnabled;
+  final String? userAgent;
 
   InitRequest({
     required this.id,
@@ -404,6 +405,7 @@ class InitRequest {
     this.androidAudioEffects = const [],
     this.darwinAudioEffects = const [],
     this.androidOffloadSchedulingEnabled,
+    this.userAgent,
   });
 
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -416,6 +418,7 @@ class InitRequest {
             .map((audioEffect) => audioEffect.toMap())
             .toList(),
         'androidOffloadSchedulingEnabled': androidOffloadSchedulingEnabled,
+        'userAgent': userAgent,
       };
 }
 
@@ -1035,6 +1038,58 @@ class AndroidLivePlaybackSpeedControlMessage {
       };
 }
 
+/// Progressive audio source options to be communicated with the platform
+/// implementation.
+class ProgressiveAudioSourceOptionsMessage {
+  final AndroidExtractorOptionsMessage? androidExtractorOptions;
+  // final DarwinAssetOptionsMessage? darwinAssetOptions;
+
+  const ProgressiveAudioSourceOptionsMessage({
+    this.androidExtractorOptions,
+    // this.darwinAssetOptions,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'androidExtractorOptions': androidExtractorOptions?.toMap(),
+        // 'darwinAssetOptions': darwinAssetOptions?.toMap(),
+      };
+}
+
+/// Options for loading audio assets on iOS/macOS to be communicated with the
+/// platform implementation.
+// class DarwinAssetOptionsMessage {
+//   final bool preferPreciseDurationAndTiming;
+//
+//   const DarwinAssetOptionsMessage({
+//     required this.preferPreciseDurationAndTiming,
+//   });
+//
+//   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+//         'preferPreciseDurationAndTiming': preferPreciseDurationAndTiming,
+//       };
+// }
+
+/// Options for extracting media files on Android to be communicated with the
+/// platform implementation.
+class AndroidExtractorOptionsMessage {
+  final bool constantBitrateSeekingEnabled;
+  final bool constantBitrateSeekingAlwaysEnabled;
+  final int mp3Flags;
+
+  const AndroidExtractorOptionsMessage({
+    required this.constantBitrateSeekingEnabled,
+    required this.constantBitrateSeekingAlwaysEnabled,
+    required this.mp3Flags,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'constantBitrateSeekingEnabled': constantBitrateSeekingEnabled,
+        'constantBitrateSeekingAlwaysEnabled':
+            constantBitrateSeekingAlwaysEnabled,
+        'mp3Flags': mp3Flags,
+      };
+}
+
 /// Information about an audio source to be communicated with the platform
 /// implementation.
 abstract class AudioSourceMessage {
@@ -1051,7 +1106,7 @@ abstract class IndexedAudioSourceMessage extends AudioSourceMessage {
   /// Since the tag type is unknown, this can only be used by platform
   /// implementations that pass by reference.
   final dynamic tag;
-  IndexedAudioSourceMessage({required String id, this.tag}) : super(id: id);
+  IndexedAudioSourceMessage({required super.id, this.tag});
 }
 
 /// Information about a URI audio source to be communicated with the platform
@@ -1061,22 +1116,25 @@ abstract class UriAudioSourceMessage extends IndexedAudioSourceMessage {
   final Map<String, String>? headers;
 
   UriAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.uri,
     this.headers,
-    dynamic tag,
-  }) : super(id: id, tag: tag);
+    super.tag,
+  });
 }
 
 /// Information about a progressive audio source to be communicated with the
 /// platform implementation.
 class ProgressiveAudioSourceMessage extends UriAudioSourceMessage {
+  final ProgressiveAudioSourceOptionsMessage? options;
+
   ProgressiveAudioSourceMessage({
-    required String id,
-    required String uri,
-    Map<String, String>? headers,
-    dynamic tag,
-  }) : super(id: id, uri: uri, headers: headers, tag: tag);
+    required super.id,
+    required super.uri,
+    super.headers,
+    super.tag,
+    this.options,
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1084,6 +1142,7 @@ class ProgressiveAudioSourceMessage extends UriAudioSourceMessage {
         'id': id,
         'uri': uri,
         'headers': headers,
+        'options': options?.toMap(),
       };
 }
 
@@ -1091,11 +1150,11 @@ class ProgressiveAudioSourceMessage extends UriAudioSourceMessage {
 /// implementation.
 class DashAudioSourceMessage extends UriAudioSourceMessage {
   DashAudioSourceMessage({
-    required String id,
-    required String uri,
-    Map<String, String>? headers,
-    dynamic tag,
-  }) : super(id: id, uri: uri, headers: headers, tag: tag);
+    required super.id,
+    required super.uri,
+    super.headers,
+    super.tag,
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1110,11 +1169,11 @@ class DashAudioSourceMessage extends UriAudioSourceMessage {
 /// implementation.
 class HlsAudioSourceMessage extends UriAudioSourceMessage {
   HlsAudioSourceMessage({
-    required String id,
-    required String uri,
-    Map<String, String>? headers,
-    dynamic tag,
-  }) : super(id: id, uri: uri, headers: headers, tag: tag);
+    required super.id,
+    required super.uri,
+    super.headers,
+    super.tag,
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1131,9 +1190,9 @@ class SilenceAudioSourceMessage extends IndexedAudioSourceMessage {
   final Duration duration;
 
   SilenceAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.duration,
-  }) : super(id: id);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1151,11 +1210,11 @@ class ConcatenatingAudioSourceMessage extends AudioSourceMessage {
   final List<int> shuffleOrder;
 
   ConcatenatingAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.children,
     required this.useLazyPreparation,
     required this.shuffleOrder,
-  }) : super(id: id);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1175,12 +1234,12 @@ class ClippingAudioSourceMessage extends IndexedAudioSourceMessage {
   final Duration? end;
 
   ClippingAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.child,
     this.start,
     this.end,
-    dynamic tag,
-  }) : super(id: id, tag: tag);
+    super.tag,
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1199,10 +1258,10 @@ class LoopingAudioSourceMessage extends AudioSourceMessage {
   final int count;
 
   LoopingAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.child,
     required this.count,
-  }) : super(id: id);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1326,9 +1385,9 @@ class AndroidLoudnessEnhancerMessage extends AudioEffectMessage {
   final double targetGain;
 
   AndroidLoudnessEnhancerMessage({
-    required bool enabled,
+    required super.enabled,
     required this.targetGain,
-  }) : super(enabled: enabled);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1418,9 +1477,9 @@ class AndroidEqualizerMessage extends AudioEffectMessage {
   final AndroidEqualizerParametersMessage? parameters;
 
   AndroidEqualizerMessage({
-    required bool enabled,
+    required super.enabled,
     required this.parameters,
-  }) : super(enabled: enabled);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{